chore: move xmodel dumping and loading code into generic files
