+++ src/ObjWriting/Game/T6/XModel/XModelDumperT6.cpp
@@ -1,0 +1,17 @@
+#include "XModelDumperT6.h"
+
+#include "Game/T6/CommonT6.h"
+#include "Game/T6/Json/JsonXModel.h"
+
+#define GAME_NAMESPACE T6
+
+#include "XModel/GenericXModelDumper.inc.h"
+
+namespace T6
+{
+    void DumpXModel(AssetDumpingContext& context, XAssetInfo<XModel>* asset)
+    {
+        DumpXModelJson(context, asset);
+        DumpXModelSurfs(context, asset);
+    }
+} // namespace T6
